backend/pyproject.toml
[project]
name = "alice-backend"
version = "0.1.0"
description = "AL\\CE — Adaptive Learning Interface for Computing & Execution (Backend)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Jays" },
]

dependencies = [
    # Core
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "websockets>=13.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-multipart>=0.0.9",
    "jsonschema>=4.0",

    # PC Automation
    "pyautogui>=0.9.54",
    "Pillow>=10.0",
    "pywinauto>=0.6.8",
    "pywin32>=306",
    "pynput>=1.7",

    # IoT
    "paho-mqtt>=2.1",

    # Web Search
    "ddgs>=9.0",
    "beautifulsoup4>=4.12",

    # Calendar
    "python-dateutil>=2.9",
    "tzdata>=2024.1",

    # Notifications (Windows)
    "winotify>=1.1",

    # Media Control (Windows)
    "pycaw>=20230407",
    "comtypes>=1.4",

    # Clipboard
    "pyperclip>=1.9",

    # News
    "feedparser>=6.0",

    # Memory (Phase 9)
    "sqlite-vec>=0.1.6",

    # MCP Client (Phase 11)
    "mcp>=1.0",

    # System Info
    "psutil>=6.0",

    # Email Assistant (Phase 15)
    "aioimaplib>=1.0.1",
    "aiosmtplib>=3.0",

    # Utils
    "loguru>=0.7",
    "watchfiles>=0.24",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
]
# STT
stt = [
    "faster-whisper>=1.1,<2",
    "numpy>=1.26",
]
# STT GPU (CUDA 12 runtime libraries)
stt-gpu = [
    "nvidia-cublas-cu12>=12.4,<13",
    "nvidia-cudnn-cu12>=9.0,<10",
]
# TTS
tts-piper = [
    "piper-tts>=1.2,<2",
]
tts-xtts = [
    "TTS>=0.22",
]
tts-kokoro = [
    "kokoro-onnx>=0.4",
    "soundfile>=0.12",
]
email = [
    "keyring>=24.0",
]
# VRAM monitoring
vram = [
    "pynvml>=12.0,<13",
]
# Voice (all-in-one: STT + TTS piper + VRAM, CPU-compatible)
voice = [
    "faster-whisper>=1.1,<2",
    "numpy>=1.26",
    "piper-tts>=1.2,<2",
    "kokoro-onnx>=0.4",
    "soundfile>=0.12",
    "pynvml>=12.0,<13",
]
# Voice with GPU acceleration (adds CUDA libraries)
voice-gpu = [
    "faster-whisper>=1.1,<2",
    "numpy>=1.26",
    "piper-tts>=1.2,<2",
    "kokoro-onnx>=0.4",
    "soundfile>=0.12",
    "pynvml>=12.0,<13",
    "nvidia-cublas-cu12>=12.4,<13",
    "nvidia-cudnn-cu12>=9.0,<10",
]
# File reading (optional)
file-reader = [
    "pdfplumber>=0.11",
    "python-docx>=1.1",
]
# Memory embedding fallback (CPU)
memory = [
    "fastembed>=0.3",
]

# tts-xtts (Coqui TTS) requires numpy<2.0 while kokoro-onnx/voice/voice-gpu
# require numpy>=2.0.2 — declare them as mutually exclusive so uv resolves
# each independently rather than failing on a combined resolution.
[tool.uv]
conflicts = [
    [
        { extra = "tts-xtts" },
        { extra = "tts-kokoro" },
    ],
    [
        { extra = "tts-xtts" },
        { extra = "voice" },
    ],
    [
        { extra = "tts-xtts" },
        { extra = "voice-gpu" },
    ],
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# The backend/ directory IS the 'backend' package but pyproject.toml lives
# inside it, so we list sub-packages explicitly for a regular install.
packages = ["core", "api", "db", "services", "plugins"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
